Use SDL_filesystem.h for data paths
* Use SDL_GetPrefPath() to get user preferences directory.
  Non-Windows & non-macOS systems will try ~/.azimuth-game first so
  existing installs aren't disrupted.
* Use SDL_GetBasePath() on macOS to retrieve app bundle resource path.
